Remove Kubernetes Helm Server
This change is inspired by a [discussion on reddit](https://www.reddit.com/r/devops/comments/itdhwr/terragruntterraform_iaac/g5f3drh?utm_source=share&utm_medium=web2x&context=3) where a user noticed this and asked if we had plans to update it. That repo is deprecated now that we've updated to Helm v3, and it might be confusing to users to see the Tiller reference there. I've also moved the Kubernetes Services Helm chart repo up to replace it.