Fix paragraph spacing in Guides (#379)
This quick PR addresses the overly tight (lack of) spacing between some paragraphs in our guides.  The way that the CSS is structured in this product leaves many ways in which inconsistency can be introduced. I'm not going to deal with the larger issues in this PR but will simply address the problem.

Co-authored-by: Josh Padnick <josh@gruntwork.io>