Update Nokogiri dependency (#393)
This PR updated the dependency for the nokogiri library. This is a library used by the `jekyll-toc` plugin and is not used directly. The reason to update the version of nokogiri is because of a security notification from Dependabot. I don't believe that there's any issues from making this update.