[APT-324] Add dev portal sign in link to navbar and footer
This adds a sign in link to the nav bar and the connect section of
the footer, which lead to the Gruntwork Developer Portal.

The navbar was using a curious combination of negative margins on
specific list items in order to achieve its layout, which made it
challenging to add another item. This change refactors the layout
into right/left navbar sections which float accordingly, and a
container that adjusts for several screen sizes.