Fix /sha endpoint
This PR hopes to fix the `/sha` endpoint. I suspect that the reason it's currently not working is that in production we won't serve files that aren't called `index.html`. If the trailing slash is missing from the `/sha` permalink definition then Jekyll will generate the file as `/sha.html`.