[APT-1373] Switch from Chargebee to Recurly for checkout.
In addition to the checkout service change, this tidies up a bit of
CSS, fixes a bug that caused incorrect subscription totals from being
shown on mobile, and replaces the $500 deposit with a "due now" total
euqal to the first month + fixed costs (i. e. RefArch), and supporting
text to indicate the ongoing monthly fee.