Attempt to fix consultingjs script
This PR attempts to fix issue loading the consulting.js script in production due to odd MIME types being reported for the file. We don't quite understand what's going on or why the MIME types appear odd. We've tried re-creating and re-formatting the file in this PR in the hopes of correcting the issue.